Machine single point calibration probing automated
The three buttons that defined the board plane had their function
altered, so that now each one of them triggers an 'exeprobe'
command to the machine and parses the serial interface return
value, storing it in the actual position of the machine. Next, the
present point is added to the calibration array of three points so
that the board may be placed along a mathematically defined plane.
Finally, a command is sent to rise the tool 10mm to clear the
board and get ready for another command from a safe position.