The "Align to XY" routine seems to be fixed
Some work was done in the comms thread, with new comm_modes being added.
In a nutshell, the origin is now set to whatever the GUI chooses, so the
set_xy_origin and set_z_origin were changed so that a set of parameters
are written to the serial port and read by the microcontroller.
Previously, the set origin modes didn't pass anything to the
microcontroller and it set its origin to its current position. The
alternative would be to send the machine to the origin of the GUI's
referential and then to set_xy_origin as was done previously. Obviously,
this would require an additional movement, which would take more time
than absolutely necessary, so these changes were made.

On a sidenote, a write_current_pos_absolute() function was added to
force the position of the machine to be the same as the GUI's. This
seems not to be necessary at this point in time, but it is being called
atm on align to xy and it works. I think it might be useful in the
future, namely when an automated probing routine is implemented to probe
across the entirety of the board.

The respective comm verifications were added, every time some parameters
are passed to the microcontroller.